Refactored the project
